Require for loops to use parenthesis
This is done in preparation for adding struct literals. With the old
syntax, the following case would be potentially annoying to parse:

    for i in RangeIterator { 0, 10 } {
        // Do something with i...
    }

Here it's difficult for the parser to tell whether we're constructing a
RangeIterator, or if we're using a variable called RangeIterator. I
don't want C-style code where the parser needs type definition feedback.
With the new syntax, we don't have any such ambiguity:

    for i in (RangeIterator { 0, 10 }) {
        // Do something with i...
    }